requirements.txt
streamlit
pandas
matplotlib
python-dotenv
psycopg2-binary
